fixup: conventions naming and move functions
